Replace simpleparse with pyparsing
Configshell uses simpleparse to parse the command line. Unfortunately,
simpleparse does not seem to be maintained anymore: the last release
was three years ago. Moreover, simpleparse is not widely used (on
Debian, only configshell depends on it).

On the other hand, pyparsing is actively maintained, widely used and
ready for Python 3. So let's use it.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@yahoo.fr>