Re-enable hack for tab-completion with nothing under the cursor
A very nice behavior is having <tab><tab> bring up options for the next
expected commandline token when nothing is underneath the cursor. There was
a workaround to get this working with the simpleparse parser, and we need
a different workaround for it to work with pyparsing.

This keeps the same technique of adding a bogus input and seeing what the
parser would categorize it as, and then setting current_token based on
that.

We also ensure current_token is defined. Without this, the call to
_dispatch_completion fails with a NameError but somebody catches the
error so things just silently don't work, without the debug msg in
_dispatch_completion for unknown current_token ever printing.

Signed-off-by: Andy Grover <agrover@redhat.com>