Fix ui_command_get() when no parameter is provided
The "get attribute" command displays an error message after the list
of attributes. For instance, in the TPG context, the following error
is displayed:

  No parameter 't' in group 'attribute'

This happens because the 'parameter' argument in ui_command_get() is
overwritten in the loop that iterates over the list of parameters to
display their value. This patch fixes the issue by renaming the
variable in the loop so that the value of the original 'parameter'
variable is preserved.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@yahoo.fr>