Fix exception on sorting
A node named "foo4" will match the regexp in sorting_keys and will sort by
("foo", 4). A node named "foo" will not match and will sort by
("foo", None).

This is incorrect because None is not orderable with integers. In the
second case, having the second element in the tuple be 0 does the
right thing.

Fixes #14

Reported-by: Arthur Lutz
Signed-off-by: Andy Grover <agrover@redhat.com>