Fix the command line completion with Python 3
The command line completion is broken with Python 3 because the
_display_completions() method raises a TypeError exception.

The root cause is a comparison between a variable named "len" and two
integer values. But because "len" is not defined at all, what is
compared is the builtin len() function and two integers. Python 3
rightly raises a TypeError exception whereas Python 2.7 does not
complain.

The fix is just to remove the "if" block containing the faulty
comparison because it always evaluates to False with Python 2.7 (len()
is always greater than "num_per_line").

This patch fixes issue #17.

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>