Remove epydoc imports and epydoc calls
Epydoc is no longer maintained and only supported for Python 2 whereas
most distributions package targetcli-fb for Python 3 (RHEL, Fedora,
Debian). As a consequence, the help messages for configshell commands
are not rendered with Epydoc on most systems and it is probably safe
to remove the Epydoc dependency entirely.

For more information, see:

  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=881554

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>