Update setup.py
Testing on a RHEL 6.10 derivative using Python 2.6.6, that re.match didn't result in a version being detected.  This fix extracted the correct version.   Appears same change would be needed in rtslib-fb and targetcli-fb as well I believe.  Don't have an opportunity to try this on a RHEL 7 though so not 100% if this would cover that version.